Crate wordweaver_schemata

Modules

v0

Macros

schemata

versions

Generates deserialize_as and SUPPORTED based on the list of versions.

Enums

Error

Constants

SUPPORTED

The versions supported by this library.

Traits

Schemata

Marker trait for schemata.

Functions

deserialize_as

Automatically picks the right version to deserialize the data and deserializes it. It returns a trait object, so you’ll need to downcast it to the right type.